#48 Fix 학교생활란 수정
#48 Fix 학교생활란 수정
