+++ src/main/java/ssg/com/a/controller/MypageController.java
@@ -14,9 +14,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 
 import ssg.com.a.dto.BbsComment;
-import ssg.com.a.dto.BbsDto;
 import ssg.com.a.dto.CalendarDto;
 import ssg.com.a.dto.CalendarParam;
+import ssg.com.a.dto.FreeBbsDto;
 import ssg.com.a.dto.FriendDto;
 import ssg.com.a.dto.MyblacklistDto;
 import ssg.com.a.service.MypageService;
@@ -49,14 +49,6 @@ public class MypageController {
 		
 		return "mypage/mychange";
 	}
-		
-	// 마이페이지 학점계산기 이동
-	@GetMapping("mygradecal.do")
-	public String mygradecal() {
-		System.out.println("MypageController mygradecal" + new Date());
-		
-		return "mypage/mygradecal";
-	}
 	
 	// 마이페이지 좋아요게시글 이동
 	@GetMapping("mylike.do")
@@ -81,7 +73,7 @@ public class MypageController {
 		
 		FriendDto login = (FriendDto)request.getSession().getAttribute("login");
 		List<BbsComment> commentList = service.getMyCommentList(login.getId());
-		List<BbsDto> writeList = service.getMyWriteList(login.getId());
+		List<FreeBbsDto> writeList = service.getMyWriteList(login.getId());
 		
 		model.addAttribute("commentList", commentList);
 		model.addAttribute("writeList", writeList);
@@ -99,78 +91,6 @@ public class MypageController {
 		model.addAttribute("bllist", bllist);
 		
 		return "mypage/myblacklist";
-	}
-			
-	// 마이페이지 학과일정달력 이동
-	@GetMapping("mycalendar.do")
-	public String mycalendar(Model model) {
-		System.out.println("MypageController mycalendar" + new Date());
-		
-		FriendDto login = (FriendDto)request.getSession().getAttribute("login");
-		Calendar cal = Calendar.getInstance();
-		cal.set(Calendar.DATE, 1);
-		
-		String syear = request.getParameter("year");
-		String smonth = request.getParameter("month");
-		
-		// 현재 연도와 월을 구한다	-> 처음 이 페이지가 실행시에 적용
-		int year = cal.get(Calendar.YEAR);
-		if (CalendarUtil.nvl(syear) == false) { // 넘어 온 파라미터 값이 있음
-			year = Integer.parseInt(syear);
-		}
-		int month = cal.get(Calendar.MONTH) +1; // 0 ~ 11 까지이므로 +1이 필요하다
-		if  (CalendarUtil.nvl(smonth) == false) {
-			month = Integer.parseInt(smonth);
-		}
-		
-		if (month < 1) {
-			month = 12;
-			year--;
-		}
-		if (month > 12) {
-			month = 1;
-			year++;
-		}
-		
-		cal.set(year, month-1, 1);
-		
-		// 요일
-		int dayOfWeek = cal.get(Calendar.DAY_OF_WEEK);
-		
-		// <<	year--
-		String pp = String.format("<a href='mycalendar.do?param=calendarList&year=%d&month=%d' style='text-decoration:none'>"
-								+ "<img src='./images/left.png' width='20px' height='20px'>"
-								+ "</a>", year-1, month);
-		
-		// <	month--
-		String p = String.format("<a href='mycalendar.do?param=calendarList&year=%d&month=%d' style='text-decoration:none'>"
-							+ "<img src='./images/prev.png' width='20px' height='20px'>"
-							+ "</a>", year, month-1);
-		
-		// >	month++
-		String n = String.format("<a href='mycalendar.do?param=calendarList&year=%d&month=%d' style='text-decoration:none'>"
-							+ "<img src='./images/next.png' width='20px' height='20px'>"
-							+ "</a>", year, month+1);	
-		
-		// >>	year++
-		String nn = String.format("<a href='mycalendar.do?param=calendarList&year=%d&month=%d' style='text-decoration:none'>"
-				+ "<img src='./images/last.png' width='20px' height='20px'>"
-				+ "</a>", year+1, month);
-		
-		CalendarParam calpa = new CalendarParam(login.getId(), year + CalendarUtil.two(month + ""), login.getMajor(), login.getCollege());
-		List<CalendarDto> list = service.getCalendarList(calpa);
-		
-		model.addAttribute("list", list);
-		model.addAttribute("pp", pp);
-		model.addAttribute("p", p);
-		model.addAttribute("nn", nn);
-		model.addAttribute("n", n);
-		model.addAttribute("year", year);
-		model.addAttribute("month", month);
-		model.addAttribute("cal", cal);
-		model.addAttribute("dayOfWeek", dayOfWeek);
-		
-		return "mycal/mycalendar";
 	}
 	
 	// TODO 블랙리스트 기능란
@@ -249,82 +169,6 @@ public class MypageController {
 		
 		return "message";
 	}
-	
-	// TODO 학과일정란
-	// 학과일정 작성하기 이동
-	@GetMapping("mycalwrite.do")
-	public String mycalwrite(Model model, String year, String month, String day) {
-		System.out.println("MypageController mycalwrite" + new Date());
-		
-		model.addAttribute("year", year);
-		model.addAttribute("month", month);
-		model.addAttribute("day", day);
-		
-		return "mycal/mycalwrite";
-	}
-	
-	// 학과일정 작성완료
-	@PostMapping("mycalwriteAf.do")
-	public String mycalwriteAf(Model model, CalendarDto dto, String date, String time) {
-		System.out.println("MypageController mycalwriteAf" + new Date());
-		
-		// 일정 작성하기 위해 데이터 변형
-		FriendDto login = (FriendDto)request.getSession().getAttribute("login");
-		date = date.replace("-", "");
-		time = time.replace(":", "");
-		String rdate = date + time;
-		dto.setRdate(rdate);
-		dto.setAuth(login.getAuth());
-		dto.setCollege(login.getCollege());
-		dto.setMajor(login.getMajor());
-		
-		System.out.println(dto.toString());
 
-		service.addCalendarWrite(dto);
-		
-		return "redirect:mycalendar.do";
-	}
-	
-	// 학과일정 자세히 보기
-	@GetMapping("mycaldetail.do")
-	public String calendarDetail(Model model, int seq) {
-		System.out.println("MypageController calendarDetail" + new Date());
-		
-		CalendarDto dto = service.calendarDetail(seq);
-		model.addAttribute("dto", dto);
-		
-		return "mycal/mycaldetail";
-	}
-	
-	// 학과일정 수정
-	@GetMapping("mycalupdate.do")
-	public String mycalupdate(Model model, int seq) {
-		System.out.println("MypageController mycalupdate" + new Date());
-		
-		CalendarDto dto = service.calendarDetail(seq);
-		model.addAttribute("dto", dto);
-		
-		return "mycal/mycalupdate";
-	}
-	
-	// 학과일정 수정완료
-	@PostMapping("mycalupdateAf.do")
-	public String mycalupdateAf(CalendarDto dto) {
-		System.out.println("MypageController mycalupdateAf" + new Date());
-		
-		service.mycalupdateAf(dto);
-		
-		return "redirect:mycalendar.do";
-	}
-	
-	// 학과일정 삭제
-	@GetMapping("mycaldelete.do")
-	public String mycaldelete(int seq) {
-		System.out.println("MypageController mycaldelete" + new Date());
-		
-		service.mycaldelete(seq);
-		
-		return "redirect:mycalendar.do";
-	}
 	
 }
