+++ src/main/java/ssg/com/a/dao/impl/MypageDaoImpl.java
@@ -8,9 +8,9 @@ import org.springframework.stereotype.Repository;
 
 import ssg.com.a.dao.MypageDao;
 import ssg.com.a.dto.BbsComment;
-import ssg.com.a.dto.BbsDto;
 import ssg.com.a.dto.CalendarDto;
 import ssg.com.a.dto.CalendarParam;
+import ssg.com.a.dto.FreeBbsDto;
 import ssg.com.a.dto.FriendDto;
 import ssg.com.a.dto.MyblacklistDto;
 
@@ -23,17 +23,12 @@ public class MypageDaoImpl implements MypageDao{
 	String ns = "mypage.";
 
 	@Override
-	public List<CalendarDto> getCalendarList(CalendarParam calpa) {
-		return session.selectList(ns + "getCalendarList", calpa);
-	}
-
-	@Override
 	public List<BbsComment> getMyCommentList(String id) {
 		return session.selectList(ns + "getMyCommentList", id);
 	}
 
 	@Override
-	public List<BbsDto> getMyWriteList(String id) {
+	public List<FreeBbsDto> getMyWriteList(String id) {
 		return session.selectList(ns + "getMyWriteList", id);
 	}
 
@@ -67,26 +62,6 @@ public class MypageDaoImpl implements MypageDao{
 	@Override
 	public int closeAccount(String id) {
 		return session.update(ns + "closeAccount", id);
-	}
-
-	@Override
-	public void addCalendarWrite(CalendarDto dto) {
-		session.insert(ns + "addCalendarWrite", dto);
-	}
-
-	@Override
-	public CalendarDto calendarDetail(int seq) {
-		return session.selectOne(ns + "calendarDetail", seq);
-	}
-
-	@Override
-	public void mycalupdateAf(CalendarDto dto) {
-		session.update(ns + "mycalupdateAf", dto);
-	}
-
-	@Override
-	public void mycaldelete(int seq) {
-		session.delete(ns + "mycaldelete", seq);
 	}
 
 }
