+++ src/main/java/ssg/com/a/service/MypageService.java
@@ -3,20 +3,17 @@ package ssg.com.a.service;
 import java.util.List;
 
 import ssg.com.a.dto.BbsComment;
-import ssg.com.a.dto.BbsDto;
 import ssg.com.a.dto.CalendarDto;
 import ssg.com.a.dto.CalendarParam;
+import ssg.com.a.dto.FreeBbsDto;
 import ssg.com.a.dto.FriendDto;
 import ssg.com.a.dto.MyblacklistDto;
 
 public interface MypageService {
 	
-	// 월별 일정 보기
-	List<CalendarDto> getCalendarList(CalendarParam calpa);
-	
 	// 내가 쓴 댓글 및 게시글 보기
 	List<BbsComment> getMyCommentList(String id);
-	List<BbsDto> getMyWriteList(String id);
+	List<FreeBbsDto> getMyWriteList(String id);
 	
 	// 블랙리스트
 	void addBlacklistId(MyblacklistDto dto);
@@ -30,9 +27,4 @@ public interface MypageService {
 	// 회원탈퇴
 	boolean closeAccount(String id);
 	
-	// 학과일정
-	void addCalendarWrite(CalendarDto dto);
-	CalendarDto calendarDetail(int seq);
-	void mycalupdateAf(CalendarDto dto);
-	void mycaldelete(int seq);
 }
