+++ src/main/java/ssg/com/a/service/serviceimpl/MypageServiceImpl.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Service;
 
 import ssg.com.a.dao.MypageDao;
 import ssg.com.a.dto.BbsComment;
-import ssg.com.a.dto.BbsDto;
 import ssg.com.a.dto.CalendarDto;
 import ssg.com.a.dto.CalendarParam;
+import ssg.com.a.dto.FreeBbsDto;
 import ssg.com.a.dto.FriendDto;
 import ssg.com.a.dto.MyblacklistDto;
 import ssg.com.a.service.MypageService;
@@ -21,17 +21,12 @@ public class MypageServiceImpl implements MypageService {
 	MypageDao dao;
 
 	@Override
-	public List<CalendarDto> getCalendarList(CalendarParam calpa) {
-		return dao.getCalendarList(calpa);
-	}
-
-	@Override
 	public List<BbsComment> getMyCommentList(String id) {
 		return dao.getMyCommentList(id);
 	}
 
 	@Override
-	public List<BbsDto> getMyWriteList(String id) {
+	public List<FreeBbsDto> getMyWriteList(String id) {
 		return dao.getMyWriteList(id);
 	}
 
@@ -64,26 +59,5 @@ public class MypageServiceImpl implements MypageService {
 	public boolean closeAccount(String id) {
 		return dao.closeAccount(id)>0?true:false;
 	}
-
-	@Override
-	public void addCalendarWrite(CalendarDto dto) {
-		dao.addCalendarWrite(dto);
-	}
-
-	@Override
-	public CalendarDto calendarDetail(int seq) {
-		return dao.calendarDetail(seq);
-	}
-
-	@Override
-	public void mycalupdateAf(CalendarDto dto) {
-		dao.mycalupdateAf(dto);
-	}
-
-	@Override
-	public void mycaldelete(int seq) {
-		dao.mycaldelete(seq);
-	}
-	
 	
 }
