+++ src/main/webapp/WEB-INF/views/collegelife/mycaldetail.jsp
