+++ src/main/webapp/WEB-INF/views/collegelife/mycalendar.jsp
@@ -23,6 +23,7 @@
 <title>마이페이지 학과일정</title>
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.7.1/jquery.min.js"></script>
+<link rel="stylesheet" href="style/css/style.css">
 <style type="text/css">
 #mymain {
 	display: flex;
@@ -43,59 +44,20 @@ a {
 </style>
 </head>
 <body>
-	<nav>
-		<table border="1">
-		<tr>
-			<th><a href="">슬갓생 </a></th>
-			<th><a href="">게시판</a></th>
-			<th><a href="">학교생활</a></th>
-			<th><a href="mymain.do">마이페이지</a></th>
-			<th><button>라이트모드</button></th>
-		</tr>
-		</table>
-	</nav>
+  <script>
+    $(document).ready( function() {
+      $("#topnav").load("topnav.do");
+      $("#leftMenu").load("collnav.do");
+    });
+  </script>
+	<nav id="topnav"></nav>
 	
-	<h1>내정보 - 학과일정</h1>
+	<header id="header">	
+		<h1>학교생활 </h1>
+	</header>
 	
-	<main id="mymain">	
-		<table id="leftMenu" border="1">
-			<tr>
-				<th><a href="mymain.do">기본정보</a></th>
-			</tr>
-			<tr>
-				<th><a href="">수강 시간표 작성</a></th>
-			</tr>	
-			<tr>
-				<th><a href="mycalendar.do">학과 일정달력</a></th>
-			</tr>	
-			<tr>
-				<th><a href="mygradecal.do">학점 계산기</a></th>
-			</tr>	
-			<tr>
-				<th><a href="">메시지</a></th>
-			</tr>	
-			<tr>
-				<th><a href="">친구 정보</a></th>
-			</tr>	
-			<tr>
-				<th><a href="mywrite.do">작성한 댓글 및 게시글</a></th>
-			</tr>	
-			<tr>
-				<th><a href="mylike.do">좋아요한 게시글</a></th>
-			</tr>
-			<tr>
-				<th><a href="myblacklist.do">블랙리스트</a></th>
-			</tr>	
-			<tr>
-				<th><a href="mychange.do">개인정보 변경</a></th>
-			</tr>	
-			<tr>
-				<th><a href="">학교 인증</a></th>
-			</tr>
-			<tr>
-				<th><a href="">회원 탈퇴</a></th>
-			</tr>
-		</table>
+	<main id="mymain">
+    <leftmenu id="leftMenu"></leftmenu>
 	
 		<calender>
 			<table border="1" >
