+++ src/main/webapp/WEB-INF/views/collegelife/mycalwrite.jsp
