+++ src/main/webapp/WEB-INF/views/mypage/myblacklist.jsp
@@ -12,6 +12,7 @@ List<MyblacklistDto> bllist = (List<MyblacklistDto>)request.getAttribute("bllist
 <meta charset="UTF-8">
 <title>마이페이지</title>
 <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.7.1/jquery.min.js"></script>
+<link rel="stylesheet" href="style/css/style.css">
 <style type="text/css">
 #mymain {
 	display: flex;
