+++ src/main/webapp/WEB-INF/views/mypage/wordlistadd.jsp
@@ -6,6 +6,8 @@
 <head>
 <meta charset="UTF-8">
 <title>마이페이지</title>
+<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.7.1/jquery.min.js"></script>
+<link rel="stylesheet" href="style/css/style.css">
 <style type="text/css">
 #mymain {
 	display: flex;
